+++ CHANGES.rst
@@ -3,6 +3,8 @@ Changelog
 
 0.1 (unreleased)
 ----------------
+* Backend portlet de video. (closes `#t_71`_).
+  [rodfersou]
 * Backend portlet de audio. (closes `#t_68`_).
   [rodfersou]
 * Backend portlet de galeria de audios. (closes `#t_65`_).
@@ -13,3 +15,4 @@ Changelog
 .. _`#t_56`: https://grupotv1.codebasehq.com/projects/secom/tickets/56
 .. _`#t_65`: https://grupotv1.codebasehq.com/projects/secom/tickets/65
 .. _`#t_68`: https://grupotv1.codebasehq.com/projects/secom/tickets/68
+.. _`#t_71`: https://grupotv1.codebasehq.com/projects/secom/tickets/71